finished 7 tables

+++ LogeenStockManagement/Controllers/ProductsController.cs
@@ -46,7 +46,8 @@ namespace LogeenStockManagement.Controllers
         [HttpPut("{id}")]
         public async Task<IActionResult> PutProduct(int id, Product product)
         {
-            if (id != product.Id)
+            //validation section1
+            if (id != product.Id || IsProductDataNotValid(product))
             {
                 return BadRequest();
             }
@@ -77,6 +78,11 @@ namespace LogeenStockManagement.Controllers
         [HttpPost]
         public async Task<ActionResult<Product>> PostProduct(Product product)
         {
+            //validation section2
+            if (IsProductDataNotValid(product))
+            {
+                return BadRequest();
+            }
             _context.Products.Add(product);
             await _context.SaveChangesAsync();
 
@@ -92,6 +98,19 @@ namespace LogeenStockManagement.Controllers
             {
                 return NotFound();
             }
+            //validation section3
+            if (
+                product.ExpiredProducts.Count>0||
+                product.ProductTransfereds.Count>0||
+                product.PurchaseProducts.Count>0||
+                product.SaleBillProducts.Count>0||
+                product.StockProducts.Count> 0
+
+                )
+            {
+                return BadRequest();
+            }
+
 
             _context.Products.Remove(product);
             await _context.SaveChangesAsync();
@@ -102,6 +121,33 @@ namespace LogeenStockManagement.Controllers
         private bool ProductExists(int id)
         {
             return _context.Products.Any(e => e.Id == id);
+        }
+        public bool IsProductDataNotValid(Product product)
+        {
+            bool categoryExisted = _context.Products.Any(ww => ww.Id == product.CategoryId);
+
+            bool Name = _context.Products.Any((p) => p.CategoryId == product.CategoryId && p.Id != product.Id);
+            if (
+                product.Name==null||
+                product.Description==null||
+                product.MiniAmount==0||
+                product.Barcode==null||
+                product.SellingPrice==0|| 
+                product.PurchasingPrice==0|| 
+                product.ExpiryPeriod==0|| 
+                product.CategoryId==0 ||
+               ! categoryExisted||
+                Name
+
+                )
+            {
+                return true;
+
+            }
+            else
+            {
+                return false;
+            }
         }
     }
 }
