+++ LogeenStockManagement/Controllers/PurchaseBillsController.cs
@@ -46,7 +46,9 @@ namespace LogeenStockManagement.Controllers
         [HttpPut("{id}")]
         public async Task<IActionResult> PutPurchaseBill(int id, PurchaseBill purchaseBill)
         {
-            if (id != purchaseBill.Id)
+            //validation section1
+
+            if (id != purchaseBill.Id || IsPurchaseBillDataNotValid(purchaseBill))
             {
                 return BadRequest();
             }
@@ -77,6 +79,12 @@ namespace LogeenStockManagement.Controllers
         [HttpPost]
         public async Task<ActionResult<PurchaseBill>> PostPurchaseBill(PurchaseBill purchaseBill)
         {
+            //validation section2
+            if (IsPurchaseBillDataNotValid(purchaseBill))
+            {
+                return BadRequest();
+
+            }
             _context.PurchaseBills.Add(purchaseBill);
             await _context.SaveChangesAsync();
 
@@ -92,6 +100,15 @@ namespace LogeenStockManagement.Controllers
             {
                 return NotFound();
             }
+            //validation section3
+            if (
+                purchaseBill.ExportPayments.Count>0||
+                purchaseBill.PurchaseProducts.Count>0||
+                purchaseBill.PurchaseReturnsBills.Count>0 
+                )
+            {
+                return BadRequest();
+            }
 
             _context.PurchaseBills.Remove(purchaseBill);
             await _context.SaveChangesAsync();
@@ -102,6 +119,48 @@ namespace LogeenStockManagement.Controllers
         private bool PurchaseBillExists(int id)
         {
             return _context.PurchaseBills.Any(e => e.Id == id);
+        }
+        public bool IsPurchaseBillDataNotValid(PurchaseBill purchaseBills)
+        {
+            bool StockExisted = _context.PurchaseBills.Any(p => p.Id == purchaseBills.StockId);
+            bool taxExisted = _context.PurchaseBills.Any(p => p.Id == purchaseBills.TaxId);
+            bool PayMethodExisted = _context.PurchaseBills.Any(p => p.Id == purchaseBills.PayMethodId);
+            bool SupplierExisted = _context.PurchaseBills.Any(p => p.Id == purchaseBills.SupplierId);
+            bool BillTypevalid = (purchaseBills.BillType == "Cash" || purchaseBills.BillType == "Debit");
+
+            //uniqe
+            bool BillCodeExisted = _context.PurchaseBills.Any((p) => p.BillCode == purchaseBills.BillCode && p.Id != purchaseBills.Id);
+            if (
+                purchaseBills.BillCode == null ||
+                !DateTime.TryParse(purchaseBills.Date.ToString(), out _) ||
+                purchaseBills.BillType == null ||
+                purchaseBills.Discount == 0 ||
+                purchaseBills.Paidup == 0 ||
+                purchaseBills.CheckNumber == null ||
+                purchaseBills.BillTotal == 0 ||
+                purchaseBills.StockId == 0 ||
+                purchaseBills.TaxId == 0 ||
+                purchaseBills.PayMethodId == 0 ||
+                purchaseBills.SupplierId == 0 ||
+                !StockExisted||
+               ! taxExisted||
+               ! PayMethodExisted||
+               ! SupplierExisted||
+                !BillTypevalid||
+               BillCodeExisted
+
+
+
+
+                )
+            {
+                return true;
+            }
+            else
+            {
+                return false;
+            }
+
         }
     }
 }
